backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
jwcrypto==1.5.6
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
qrcode==8.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
